% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deem_cluster.R
\name{deem_cluster}
\alias{deem_cluster}
\title{Generate simulation parameters}
\usage{
deem_cluster(
  X,
  nclass,
  niter = 100,
  lambda = NULL,
  dfmax = n,
  pmax = nvars,
  pf = rep(1, nvars),
  eps = 1e-04,
  maxit = 1e+05,
  sml = 1e-06,
  verbose = FALSE,
  perturb = NULL,
  ceps = 0.1,
  initial = FALSE,
  vec_x = NULL,
  utrue,
  vtrue,
  mutrue
)
}
\arguments{
\item{X}{aa}

\item{nclass}{aa}

\item{niter}{aa}

\item{lambda}{aa}

\item{dfmax}{aa}

\item{pmax}{aa}

\item{pf}{aa}

\item{eps}{aa}

\item{maxit}{aa}

\item{sml}{aa}

\item{verbose}{aa}

\item{perturb}{aa}

\item{ceps}{aa}

\item{initial}{aa}

\item{vec_x}{aa}

\item{utrue}{aa}

\item{vtrue}{aa}

\item{mutrue}{aa}
}
\value{
List of parameters
}
\description{
Generate simulation parameters
}
\examples{
A = array(rnorm(24),dim=c(2,3,4))
}
